refactor : 각 말들의 이동 방향을 반환해주는 메서드들 Direction 안으로 이동
