+++ src/main/java/chess/domain/direction/Direction.java
@@ -1,12 +1,15 @@
 package chess.domain.direction;
 
+import java.util.Arrays;
+import java.util.List;
+
 public enum Direction {
 
     UP(0, 1),
     DOWN(0, -1),
     RIGHT(1, 0),
     LEFT(-1, 0),
-    UP_RIGHT(1,1),
+    UP_RIGHT(1, 1),
     UP_LEFT(-1, 1),
     DOWN_RIGHT(1, -1),
     DOWN_LEFT(-1, -1),
@@ -19,8 +22,8 @@ public enum Direction {
     KNIGHT_LEFT_UP(-2, 1),
     KNIGHT_LEFT_DOWN(-2, -1);
 
-    private int x;
-    private int y;
+    private final int x;
+    private final int y;
 
     Direction(int x, int y) {
         this.x = x;
@@ -33,5 +36,65 @@ public enum Direction {
 
     public int getY() {
         return y;
+    }
+
+    public List<Direction> diagonalDirections() {
+        return Arrays.asList(
+                UP_RIGHT,
+                UP_LEFT,
+                DOWN_RIGHT,
+                DOWN_LEFT
+        );
+    }
+
+    public List<Direction> straightDirections() {
+        return Arrays.asList(
+                UP,
+                DOWN,
+                RIGHT,
+                LEFT
+        );
+    }
+
+    public List<Direction> allDirections() {
+        return Arrays.asList(
+                UP_RIGHT,
+                UP_LEFT,
+                DOWN_RIGHT,
+                DOWN_LEFT,
+                UP,
+                DOWN,
+                RIGHT,
+                LEFT
+        );
+    }
+
+    public List<Direction> whitePawnDirections() {
+        return Arrays.asList(
+                UP,
+                UP_RIGHT,
+                UP_LEFT
+        );
+    }
+
+    public List<Direction> blackPawnDirections() {
+        return Arrays.asList(
+                DOWN,
+                DOWN_RIGHT,
+                DOWN_LEFT
+        );
+    }
+
+    public List<Direction> knightDrections() {
+        return Arrays.asList(
+                KNIGHT_UP_RIGHT,
+                KNIGHT_UP_LEFT,
+                KNIGHT_RIGHT_UP,
+                KNIGHT_RIGHT_DOWN,
+                KNIGHT_DOWN_RIGHT,
+                KNIGHT_DOWN_LEFT,
+                KNIGHT_LEFT_UP,
+                KNIGHT_LEFT_DOWN
+        );
     }
 }
