--- src/main/java/chess/domain/direction/StraightDirections.java
@@ -1,18 +1,0 @@
-package chess.domain.direction;
-
-import java.util.Arrays;
-import java.util.List;
-
-import static chess.domain.direction.Direction.*;
-
-public class StraightDirections {
-
-    public List<Direction> straightDirections() {
-        return Arrays.asList(
-                UP,
-                DOWN,
-                RIGHT,
-                LEFT
-        );
-    }
-}
