--- src/main/java/chess/domain/direction/WhitePawnDirections.java
@@ -1,19 +1,0 @@
-package chess.domain.direction;
-
-import java.util.Arrays;
-import java.util.List;
-
-import static chess.domain.direction.Direction.*;
-import static chess.domain.direction.Direction.DOWN_LEFT;
-
-public class WhitePawnDirections {
-
-    public List<Direction> whitePawnDirections() {
-        return Arrays.asList(
-                UP_RIGHT,
-                UP_LEFT,
-                DOWN_RIGHT,
-                DOWN_LEFT
-        );
-    }
-}
